tidy up compiler warnings

+++ IRCCloud/Classes/NotificationsDataSource.m
@@ -44,7 +44,7 @@
                 [[NSUserDefaults standardUserDefaults] removeObjectForKey:@"cacheVersion"];
             }
         }
-        CLS_LOG(@"NotificationsDataSource initialized with %i items from cache", _notifications.count);
+        CLS_LOG(@"NotificationsDataSource initialized with %lu items from cache", (unsigned long)_notifications.count);
     }
     return self;
 }
@@ -97,7 +97,7 @@
     }
     //[[UIApplication sharedApplication] presentLocalNotificationNow:n];
 #endif
-    CLS_LOG(@"Got notification for bid%i eid%.0f, badge count is now %i", bid, eid, _notifications.count);
+    CLS_LOG(@"Got notification for bid%i eid%.0f, badge count is now %lu", bid, eid, (unsigned long)_notifications.count);
 }
 
 -(void)removeNotificationsForBID:(int)bid olderThan:(NSTimeInterval)eid {
@@ -115,7 +115,7 @@
             [_notifications removeObjectForKey:@(bid)];
     }
 #endif
-    CLS_LOG(@"Cleared notifications for bid%i older than eid%.0f, badge count is now %i", bid, eid, _notifications.count);
+    CLS_LOG(@"Cleared notifications for bid%i older than eid%.0f, badge count is now %lu", bid, eid, (unsigned long)_notifications.count);
 }
 
 -(UILocalNotification *)getNotification:(NSTimeInterval)eid bid:(int)bid {
